envoy: separate logic to generate caches
This will allow us to test that logic separately from the code that
generates the snapshot config, updates the status of the ingress, etc.
